docrequirements.txt
psycopg2-binary>=2.9.9
markdown>=3.5.1
python-dateutil>=2.8.2
PyYAML>=6.0.1
SQLAlchemy>=2.0.25
alembic>=1.13.1
click>=8.1.7
rich>=13.7.0
watchdog>=3.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3
aiofiles>=23.2.1
python-multipart>=0.0.6
pydantic>=2.5.3
starlette>=0.36.1
httpx>=0.26.0
websockets>=12.0
networkx>=3.2.1
matplotlib>=3.8.2
beautifulsoup4>=4.12.3
numpy>=1.26.3
scipy>=1.12.0
pandas>=2.2.0
spacy>=3.7.2
textstat>=0.7.3
transformers>=4.36.2
nltk>=3.8.1
torch>=2.1.2
sentencepiece>=0.1.99
protobuf>=4.25.1 